On default to SRI if asset is from manifest
Previously we always tried to add SRI if `#asset_integrity` returned
something. With the new passthrough option we can be called with assets
from e.g. sprockets. Calling `#asset_integrity` with such assets can
return present but invalid hashs.

This commit changes the behavior to only default to adding available SRI
if the asset is from the manifest and not from an upstream plugin. This
avoids defaulting to adding SRI for e.g. sprocket assets.