Remove onboot manifest check
There just isn't any reliable way to detect if we're running in a server
process only to raise on an missing manifest file. This code always runs
in some e.g. rake tasks to or not all if conditions are too strict.
Rails and Railties to not seems to have a callback for `rails server`
only.

The function was broken anyway as with using globs a missing file will
just not be matched by `Dir.glob`. This hasn't been worked since 2.0.