fix!: Avoid issues with `config.assets` in apps without sprockets
Rails applications without sprockets (default on Rails 7+) might throw
errors, as some Rails components assume that, if `config.assets` exists,
sprockets is active and used. Booting such an application can result in
errors such as:

    NoMethodError: undefined method `+' for nil:NilClass

    Rails.application.config.assets.precompile += %w( activestorage activestorage.esm )
                                    ^

This CL changes the config tree to configure this plugin to
`config.assets_manifest`. This will not conflict with sprockets-rails.
It also reverts the change to accepts multiple manifests in
`config.assets.manifests`, which was added because
`config.assets.manifest` already did conflict with sprockets. The new
manifest will only accept a single file and error when that file cannot
be loaded.